#227 fix: repair doctests broken by ruff unused-import removal
- statevector.py: REAL was removed as unused by ruff, but the doctest
  used it as dtype — replace with float (equivalent for the test's purpose)
- puredephasing.py: module-level Aggregate import was replaced by lazy
  import, but doctest uses Aggregate — add explicit import to doctest

+++ quantarhei/qm/hilbertspace/statevector.py
@@ -27,7 +27,7 @@ class StateVector(BasisManaged):
     2
 
 
-    >>> vec = numpy.zeros((1,3), dtype=REAL)
+    >>> vec = numpy.zeros((1,3), dtype=float)
     >>> psi = StateVector(data=vec)
     Traceback (most recent call last):
     ...
